fix: adjust default values for requests and scaling targets (#17)
An scaling target over 100% can be unreliable in case of CPU starvation
caused by other workloads. Instead we should to keep the value <= 100
and adjust requests accordingly.

This also increases the LMS and CMS memory requests to 800Mi considering
that each uWSGI worker consumes around 350Mi and tutor defaults to two
of them.